Update all packages
https://github.com/advisories/GHSA-5crp-9r3c-p9vr
