+++ Jsonize.Benchmarks/Jsonize.Benchmarks.csproj
@@ -18,6 +18,7 @@
     <ItemGroup>
       <PackageReference Include="BenchmarkDotNet" Version="0.13.1" />
       <PackageReference Include="JackWFinlay.Jsonize" Version="1.0.9" />
+      <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     </ItemGroup>
 
 </Project>
