+++ Jsonize.DependencyInjection/Jsonize.DependencyInjection.csproj
@@ -1,0 +1,20 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\Jsonize.Abstractions\Jsonize.Abstractions.csproj" />
+      <ProjectReference Include="..\Jsonize.Parser\Jsonize.Parser.csproj" />
+      <ProjectReference Include="..\Jsonize.Serializer.Json.Net\Jsonize.Serializer.Json.Net.csproj" />
+      <ProjectReference Include="..\Jsonize\Jsonize.csproj" />
+    </ItemGroup>
+
+</Project>
