+++ Jsonize.Parser/Jsonize.Parser.csproj
@@ -30,12 +30,9 @@
 
     <ItemGroup>
       <PackageReference Include="AngleSharp" Version="1.0.0-alpha-844" />
-      <PackageReference Include="AngleSharp.Css" Version="0.16.3" />
+      <PackageReference Include="AngleSharp.Css" Version="0.16.4" />
       <PackageReference Include="AngleSharp.Js" Version="0.15.0" />
-    </ItemGroup>
-
-    <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' ">
-        <PackageReference Include="System.Memory" Version="4.5.4" />
+      <PackageReference Include="System.Memory" Version="4.5.5" />
     </ItemGroup>
 
 </Project>
