+++ Jsonize.Serializer/Jsonize.Serializer.csproj
@@ -28,8 +28,8 @@
       <ProjectReference Include="..\Jsonize.Abstractions\Jsonize.Abstractions.csproj" />
     </ItemGroup>
 
-    <ItemGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' ">
-        <PackageReference Include="System.Text.Json" Version="6.0.1" />
+    <ItemGroup>
+      <PackageReference Include="System.Text.Json" Version="6.0.5" />
     </ItemGroup>
 
 </Project>
