+++ Jsonize/Jsonize.csproj
@@ -31,8 +31,4 @@
         <ProjectReference Include="..\Jsonize.Abstractions\Jsonize.Abstractions.csproj" />
     </ItemGroup>
 
-    <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="6.0.0" />
-    </ItemGroup>
-
 </Project>
